chore: use i64 instead of i32 & usize

+++ src/abi/completion/chat.rs
@@ -36,7 +36,7 @@ pub struct ChatRequest {
     /// Controls how long the model will stay loaded into memory following the request
     /// (default: 5m).
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub keep_alive: Option<usize>,
+    pub keep_alive: Option<i64>,
 }
 
 #[derive(Debug, Clone, Deserialize)]
@@ -55,22 +55,22 @@ pub struct ChatResponse {
     pub done: bool,
 
     /// Time spent generating the response.
-    pub total_duration: Option<usize>,
+    pub total_duration: Option<i64>,
 
     /// Time spent in nanoseconds loading the model.
-    pub load_duration: Option<usize>,
+    pub load_duration: Option<i64>,
 
     /// Number of tokens in the prompt.
-    pub prompt_eval_count: Option<usize>,
+    pub prompt_eval_count: Option<i64>,
 
     /// Time in nanoseconds spent generating the response
-    pub prompt_eval_duration: Option<usize>,
+    pub prompt_eval_duration: Option<i64>,
 
     /// Number of tokens in the response.
-    pub eval_count: Option<usize>,
+    pub eval_count: Option<i64>,
 
     /// Time in nanoseconds spent generating the response.
-    pub eval_duration: Option<usize>,
+    pub eval_duration: Option<i64>,
 }
 
 #[derive(Debug, Clone, Deserialize)]
